requirements.txt
# Binance API and Exchange Connectivity
python-binance==1.0.19
ccxt==4.2.25

# WebSocket and Async HTTP
websockets==12.0
aiohttp==3.9.3

# Data Processing and Analysis
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9  # PostgreSQL adapter
alembic==1.13.1  # Database migrations

# Environment and Configuration
python-dotenv==1.0.1

# HTTP Requests
requests==2.31.0
